Add files via upload
I few things 2 note, this must be in a folder that has access to pygame, I just put it in the pygame examples folder because anything in that folder obviously has access to pygame. Also something to note about the game itself, this one is programmed so that the very first position you can CAN be a mine.